detail tv clear

+++ lib/data/models/tv_series_detail_model.dart
@@ -14,9 +14,9 @@ class TvSeriesDetailResponse extends Equatable {
   final int numberOfSeasons;
   final String originalName;
   final String overview;
-  final int popularity;
+  final double popularity;
   final String posterPath;
-  final int voteAverage;
+  final double voteAverage;
   final int voteCount;
 
   TvSeriesDetailResponse(
